Moved step definitions which were in CommonSteps.cs into the library Having the definitions in the .tt file means the feature files get the correct syntax highlighting
