+++ Solutions/Endjin.SpecFlow.Selenium.Sample.Specs.Spa/WebsiteUnderTest/GenerateWebTests.tt
@@ -21,6 +21,8 @@ namespace Endjin.SpecFlow.Selenium.Framework
     using Endjin.SpecFlow.Selenium.Framework.Environment;
     using Endjin.SpecFlow.Selenium.Framework.Navigation;
 
+	using Should;
+
     using TechTalk.SpecFlow;
 
     #endregion
@@ -31,7 +33,7 @@ namespace Endjin.SpecFlow.Selenium.Framework
         [BeforeFeature]
         public static void FeatureSetup()
         {
-            if (TestEnvironment.Current.IsLocal && TestEnvironment.Current.AutoStartIIS)
+			if (TestEnvironment.Current.IsLocal && TestEnvironment.Current.AutoStartLocalIIS)
             {
                 TestEnvironment.Current.StartWebsite();
             }
@@ -61,6 +63,22 @@ namespace Endjin.SpecFlow.Selenium.Framework
         public static void ScenarioTeardown()
         {
             BrowserTest.Teardown();
+        }
+    }
+
+	[Binding]
+    public class CommonSteps
+    {
+        [Given(@"I go to the (.*) page")]
+        public void IGoToThePage(string page)
+        {
+            Navigator.Browser.GoToPageByName(page);
+        }
+
+        [Then(@"I should be taken to the (.*) page")]
+        public void IShouldBeOnThePage(string page)
+        {
+            Navigator.Browser.IsOnPage(page).ShouldBeTrue();
         }
     }
 }
